research_explorer_project/requirements.txt
crewai[google-genai]
openai
python-dotenv
